Renamed uge_check --> node_search
I felt that uge_check was not a sufficient name for this script. Renaming it to node_search explains better what it is that the script is trying to accomplish, and it is easy to say, understand, and remember. Readme was updated to reflect this change, as was the man page.