Added memory functionality to running jobs. Added bash script.
Now when searching for a user and displaying said user's job
information, the maximum amount of memory used is found and displayed.
This means that node_seach.py can also be used to monitor memory usage
with jobs which can be helpful when several different users are running
on the same node.
Loading the Python 3 module can be a pain sometimes, so to streamline
the use of node_search.py the node_search.sh script was created to use
smoke and mirrors to let the Python 3 script run without needing to load
the Python 3 module. Inside the bash script things are specific to CRC
use, so if the systems change or anythings moves this bash script will
need to also be changed. The full path of the Python 3 script will also
need to be properly specified within this bash script.