Added --all capability. Cleaned up user's --details output.
The --all capability works now, going through and finding each node without duplication due to interconnection of hostgroups. The output of a user's --details is now less cluttered and more appealing. It displays all jobs running on nodes the specified user's jobs are running on. Also display's virtual memory. Slightly changed the amount of characters the screenwidth is when using --visual or -v, from 120 to 119 and only displaying 39 cores on a line instead of 40. This is done as when using split panes in Tmux it looked funky, so to have it line up better it uses one less core per line.