Changed detailed view for user's jobs.
The detailed view for a user's job will now show instead all of the processes which resulted from their jobs instead of only their jobs as recognized in UGE. Each process will also have their respective amount of resident memory displayed. All information for the processes are grabbed from xymon with their hostname passed in dynamically. This means it may be exactly up to date, but all of the information is true to within the last 2-3 minutes. This allows information to be obtained without needing to land on or probe the node, as UGE doesn't track memory very well.