Added pretty-fied memory for -u JohnDoe --details. Adjusted man page.
The memory printed from the --details flag for a user's jobs will now show easily read memory units. The units shown are; KB, MB, and GB. TB is not reflected as very few machines have mroe than 1 TB of memory anyways. Perhaps in 5 years that will change. The manual page was updated to reflect the change in --details for a user's jobs showing the processes now instead of only the jobs through UGE.