Fixed #2. Correction to #2 as follows.
GPU jobs were handled correctly UNLESS they were running on the last node reported in qstat. The last node does not have the line of separation that the script searchers for to properly deliminate different nodes in qstat's output. This would cause the last node to grab all of the pending jobs and screw up formatting etc and toss the last node into the pending jobs list and have it be lost for eternity. To mend this, the script now checks if a node is found within the pending jobs and parses it out to be added to the node list.