Changed formatting of non --detailed jobs. Fixed total mem.
Total memory could possibly be off due to an incorrect division in cleanMem. This has been fixed and so proper total memory is now displayed. The formatting of the non detailed output of searching for a user's jobs now closely represents the detailed output in terms of formatting. The total amount of memory has been added to the non detailed display as well.