Fixed issue with -uf and false users. Better user checking altogether.
Node_search now utilizes the pwd module specifically getpwnam to search /etc/passwd to see if a user exists instead of going through qconf -suserl. This way, if the user does indeed exist but hasn't submitted a job in awhile or whatever the case is, they will still be recognized as a real user instead of a nonsense of error of they may or may not exist. This also solves the issue of -uf showing that any import to it has access to the general_access and debug queues. The man page was updated to reflect this under BUGS.