Changed the way module are handled. Fixed free core count.
The module are now easier to handle as they are loaded using the init script and the actual module command now. This is easier to maintain, when it's time to upgrade python3 module simply edit the one line within node_search.sh to include the newest python module to run the node_search python script. There was a problem in the way free cores were being reported while searching for a host groups utilization. The free cores were not taking the disabled cores into account and were reporting disabled cores as free, which is not true as those cores cannot be used for jobs any longer for whatever reason.